requirements.txt
pypdf==4.3.1
reportlab==4.2.5
Pillow==10.4.0
Typer==0.12.5
rich==13.9.2
PySide6==6.9.3
PySide6-Addons==6.9.3
